requirements.txt
python-telegram-bot[socks]==12.2.0
ephem==3.7.7.0

